+++ pkg/distributor/distributor_test.go
@@ -100,8 +100,6 @@ func TestConfig_Validate(t *testing.T) {
 func TestDistributor_Push(t *testing.T) {
 	// Metrics to assert on.
 	lastSeenTimestamp := "cortex_distributor_latest_seen_sample_timestamp_seconds"
-	distributorAppend := "cortex_distributor_ingester_appends_total"
-	distributorAppendFailure := "cortex_distributor_ingester_append_failures_total"
 	distributorSampleDelay := "cortex_distributor_sample_delay_seconds"
 	ctx := user.InjectOrgID(context.Background(), "user")
 
@@ -198,17 +196,9 @@ func TestDistributor_Push(t *testing.T) {
 			happyIngesters:   2,
 			samples:          samplesIn{num: 1, startTimestampMs: now.UnixMilli() - 80000*1000}, // 80k seconds old
 			metadata:         0,
-			metricNames:      []string{distributorAppend, distributorAppendFailure, distributorSampleDelay},
+			metricNames:      []string{distributorSampleDelay},
 			expectedResponse: emptyResponse,
 			expectedMetrics: `
-				# HELP cortex_distributor_ingester_append_failures_total The total number of failed batch appends sent to ingesters.
-				# TYPE cortex_distributor_ingester_append_failures_total counter
-				cortex_distributor_ingester_append_failures_total{ingester="2",type="samples"} 1
-				# HELP cortex_distributor_ingester_appends_total The total number of batch appends sent to ingesters.
-				# TYPE cortex_distributor_ingester_appends_total counter
-				cortex_distributor_ingester_appends_total{ingester="0",type="samples"} 1
-				cortex_distributor_ingester_appends_total{ingester="1",type="samples"} 1
-				cortex_distributor_ingester_appends_total{ingester="2",type="samples"} 1
 				# HELP cortex_distributor_sample_delay_seconds Number of seconds by which a sample came in late wrt wallclock.
 				# TYPE cortex_distributor_sample_delay_seconds histogram
 				cortex_distributor_sample_delay_seconds_bucket{le="30"} 0
@@ -233,17 +223,9 @@ func TestDistributor_Push(t *testing.T) {
 			happyIngesters:   2,
 			samples:          samplesIn{num: 1, startTimestampMs: now.UnixMilli() - 1000}, // 1 second old
 			metadata:         0,
-			metricNames:      []string{distributorAppend, distributorAppendFailure, distributorSampleDelay},
+			metricNames:      []string{distributorSampleDelay},
 			expectedResponse: emptyResponse,
 			expectedMetrics: `
-				# HELP cortex_distributor_ingester_append_failures_total The total number of failed batch appends sent to ingesters.
-				# TYPE cortex_distributor_ingester_append_failures_total counter
-				cortex_distributor_ingester_append_failures_total{ingester="2",type="samples"} 1
-				# HELP cortex_distributor_ingester_appends_total The total number of batch appends sent to ingesters.
-				# TYPE cortex_distributor_ingester_appends_total counter
-				cortex_distributor_ingester_appends_total{ingester="0",type="samples"} 1
-				cortex_distributor_ingester_appends_total{ingester="1",type="samples"} 1
-				cortex_distributor_ingester_appends_total{ingester="2",type="samples"} 1
 				# HELP cortex_distributor_sample_delay_seconds Number of seconds by which a sample came in late wrt wallclock.
 				# TYPE cortex_distributor_sample_delay_seconds histogram
 				cortex_distributor_sample_delay_seconds_bucket{le="30"} 1
@@ -268,17 +250,9 @@ func TestDistributor_Push(t *testing.T) {
 			happyIngesters:   2,
 			samples:          samplesIn{num: 0, startTimestampMs: 123456789000},
 			metadata:         1,
-			metricNames:      []string{distributorAppend, distributorAppendFailure, distributorSampleDelay},
+			metricNames:      []string{distributorSampleDelay},
 			expectedResponse: emptyResponse,
 			expectedMetrics: `
-				# HELP cortex_distributor_ingester_append_failures_total The total number of failed batch appends sent to ingesters.
-				# TYPE cortex_distributor_ingester_append_failures_total counter
-				cortex_distributor_ingester_append_failures_total{ingester="2",type="metadata"} 1
-				# HELP cortex_distributor_ingester_appends_total The total number of batch appends sent to ingesters.
-				# TYPE cortex_distributor_ingester_appends_total counter
-				cortex_distributor_ingester_appends_total{ingester="0",type="metadata"} 1
-				cortex_distributor_ingester_appends_total{ingester="1",type="metadata"} 1
-				cortex_distributor_ingester_appends_total{ingester="2",type="metadata"} 1
 				# HELP cortex_distributor_sample_delay_seconds Number of seconds by which a sample came in late wrt wallclock.
 				# TYPE cortex_distributor_sample_delay_seconds histogram
 				cortex_distributor_sample_delay_seconds_bucket{le="30"} 0